rfct: add menu to notification items
